chore(plugin-flow-builder): rename hubtype-models to flow-builder-models
