+++ packages/botonic-plugin-flow-builder/src/content-fields/text.tsx
@@ -1,7 +1,7 @@
 import { Text } from '@botonic/react'
 import React from 'react'
 
-import { ButtonStyle, HtTextNode } from '../hubtype-models'
+import { ButtonStyle, HtTextNode } from '../flow-builder-models'
 import { FlowButton } from './button'
 import { ContentFieldsBase } from './content-base'
